feat: add WorkflowEvaluationsPanel component for managing workflow evaluations with features like search, pagination, and evaluation actions into a tab inside workflow page
